优化 包含英文的字母的拼音转换 新增 泛型搜索类型 优化 添加对字典以及KeyValuePair支持 优化 权重计算

+++ Pinyin.NET/SearchResults.cs
@@ -2,12 +2,12 @@
 public struct SearchResults 
 {
     public object Source { get; set; }
-    public int Weight { get; set; }
+    public double Weight { get; set; }
     public bool[] CharMatchResults { get; set; }
 }
 public struct SearchResults<T> 
 {
     public T Source { get; set; }
-    public int Weight { get; set; }
+    public double Weight { get; set; }
     public bool[] CharMatchResults { get; set; }
 }
